requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.5
django-ckeditor==6.1.0
django-cloudinary-storage==0.3.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
dnspython==1.16.0
eventlet==0.31.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
kombu==5.1.0
Pillow==8.3.1
prompt-toolkit==3.0.19

python-decouple==3.4
python-dotenv==0.19.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
